convert GuildMember
