+++ src/structures/GuildMember.js
@@ -335,7 +335,17 @@ class GuildMember {
    * @returns {Promise<GuildMember>}
    */
   edit(data) {
-    return this.client.rest.methods.updateGuildMember(this, data);
+    if (data.channel) {
+      data.channel_id = this.client.resolver.resolveChannel(data.channel).id;
+      data.channel = null;
+    }
+    if (data.roles) data.roles = data.roles.map(role => role instanceof Role ? role.id : role);
+    let endpoint = this.client.rest.api.guilds(this.guild.id).members(this.id);
+    if (this.user.id === this.client.user.id) {
+      const keys = Object.keys(data);
+      if (keys.length === 1 && keys[0] === 'nick') endpoint = endpoint.nickname;
+    }
+    return endpoint.patch({ data }).then(newData => this.guild._updateMember(this, newData).mem);
   }
 
   /**
@@ -382,7 +392,10 @@ class GuildMember {
   addRole(role) {
     if (!(role instanceof Role)) role = this.guild.roles.get(role);
     if (!role) return Promise.reject(new TypeError('Supplied parameter was neither a Role nor a Snowflake.'));
-    return this.client.rest.methods.addMemberRole(this, role);
+    if (this._roles.includes(role.id)) return Promise.resolve(this);
+    return this.client.rest.api.guilds(this.guild.id).members(this.user.id).roles(role.id)
+      .put()
+      .then(() => this);
   }
 
   /**
@@ -394,9 +407,9 @@ class GuildMember {
     let allRoles;
     if (roles instanceof Collection) {
       allRoles = this._roles.slice();
-      for (const role of roles.values()) allRoles.push(role.id);
+      for (const role of roles.values()) allRoles.push(role.id ? role.id : role);
     } else {
-      allRoles = this._roles.concat(roles);
+      allRoles = this._roles.concat(roles.map(r => r.id ? r.id : r));
     }
     return this.edit({ roles: allRoles });
   }
@@ -409,7 +422,9 @@ class GuildMember {
   removeRole(role) {
     if (!(role instanceof Role)) role = this.guild.roles.get(role);
     if (!role) return Promise.reject(new TypeError('Supplied parameter was neither a Role nor a Snowflake.'));
-    return this.client.rest.methods.removeMemberRole(this, role);
+    return this.client.rest.api.guilds(this.guild.id).members(this.user.id).roles(role.id)
+      .delete()
+      .then(() => this);
   }
 
   /**
@@ -464,7 +479,13 @@ class GuildMember {
    * @returns {Promise<GuildMember>}
    */
   kick(reason) {
-    return this.client.rest.methods.kickGuildMember(this.guild, this, reason);
+    return this.client.rest.api.guilds(this.guild.id).members(this.user.id).delete({ reason })
+    .then(() =>
+      this.client.actions.GuildMemberRemove.handle({
+        guild_id: this.guild.id,
+        user: this.user,
+      }).member
+    );
   }
 
   /**
